This is my first check in.
Well this is my first check in of the Kata.  I kinda realize now I should have done this a while ago so you can see the progression.  At the time I was more concerned with the Kata.  while checking in the folder I see that the OBJ file was included.  TFS leaves these out by default. sorry for the inclusion.