Moved data validation to the HoursCalculator Class
I felt that it was better to have the validation in the HoursCalculator class.  it just makes good sense and makes the class able to stand on it's own.  Also made variables out of the earliest start time, latest start time, and military timespans.  this way if those were to change they can be done in one place.